Add files via upload
This connector helps you to authorize the user and get the user history from Uber account.
 
The connector makes use of oAUTH2 authorization. In order make use of this authorization  developer needs to create the APPs in the https://developer.uber.com/. Once you create the app you can get the client key and client secret.   We get the security token for each authorized call which will be used to get the user history details.